feat: added an "out" task/activity - send email
NOTE: still in concept - to think about and discuss

"out" activities send data from Kotive to the service
"in" activities receive data from the service (either from a post, poll
or webhook)
Each JSON file represents a single task.

In Flow, we’ll use the mandrill/mandrill.png as icon and the name in
config.json for the Activity. When adding the new activity the user
sees a list of available functions (both IN and OUT) for that service.